requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
axe-selenium-python==2.1.6
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.30.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
distlib==0.3.8
distro==1.9.0
dj-database-url==2.3.0
Django==5.0.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastapi==0.115.6
filelock==3.14.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
msrest==0.7.1
mysql-connector-python==8.4.0
numpy==2.1.1
oauthlib==3.2.2
openai==1.57.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.0.0
platformdirs==4.2.2
playwright==1.49.1
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==11.0.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyppeteer==2.0.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.41.3
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
uvicorn==0.32.1
virtualenv==20.26.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.21.0
